+++ views/oops/noaccounts.php
@@ -1,15 +1,13 @@
-	<form action="<?php echo BASE_URL;?>users/validate" method="post">
 	<div class="container spacer centered">
 		<div class="row">
 			<div><i class="logo fa fa-thumbs-o-down fa-5x"></i></div>
 			<h2>Oops!</h2>
 			<p>Sorry, no accounts found as yet. You will have to wait till the account owner activates a few accounts for you.</p>
 		</div>
-	
+
 		<div class="spacer">
-			<a href="<?php echo BASE_URL;?>users/inform"><span class="btn btn-border">Click here to inform the account owner</span></a>
+			<a href="<?php echo BASE_URL;?>"><span class="btn btn-border">Take me back to the start</span></a>
 		</div>
 
 
 	</div>
-	</form>